+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Unreleased
 
 * Change the debug message in `debugAnnotationProcessingRootElements()` to be less prone to confusion.
+* Change the `processTypeElements(...)` and `performAction(...)` methods to accept a label to use when emitting debug messages.
 
 ### [v0.52](https://github.com/realityforge/proton/tree/v0.52) (2021-11-10) · [Full Changelog](https://github.com/realityforge/proton/compare/v0.51...v0.52)